Crate flowistry_pdg

Modules

pdg 🔒

The representation of the PDG.

rustc 🔒

rustc_impls 🔒

rustc_portable

Exports either rustc identifiers or their proxies depending on whether the rustc feature is enabled.

rustc_proxies

Proxies for Rustc types used within the PDG.

Structs

CallString

A location within the global call-graph.

GlobalLocation

A RichLocation within a specific point in a codebase.

SimpleSizedAllocativeWrapper

A wrapper that guarantees to be the same size as T and implements the Allocative trait.

Enums

RichLocation

Extends a MIR body’s Location with Start (before the first instruction) and End (after all returns).

SourceUse

Additional information about the source of data.

TargetUse

Additional information about this mutation.

Functions

allocative_visit_intern_t

allocative_visit_map_coerce_key

A function that is fit to be handed to #[allocative(visit = "...")] for a map where the key does not implement Allocative, but also has no attached heap data.

allocative_visit_simple_sized

A function fit to be handed to #[allocative(visit = "...")], if the type does not have any attached heap data.