flowistry_pdg::rustc::mir
Enum LocalInfo

pub enum LocalInfo<'tcx> {
    User(BindingForm<'tcx>),
    StaticRef {
        def_id: DefId,
        is_thread_local: bool,
    },
    ConstRef {
        def_id: DefId,
    },
    AggregateTemp,
    BlockTailTemp(BlockTailInfo),
    IfThenRescopeTemp {
        if_then: HirId,
    },
    DerefTemp,
    FakeBorrow,
    Boring,
}

Extra information about a some locals thatโs used for diagnostics and for classifying variables into local variables, statics, etc, which is needed e.g. for borrow checking.

Not used for non-StaticRef temporaries, the return place, or anonymous function parameters.

Variants

User(BindingForm<'tcx>)

A user-defined local variable or function parameter

The BindingForm is solely used for local diagnostics when generating warnings/errors when compiling the current crate, and therefore it need not be visible across crates.

StaticRef

A temporary created that references the static with the given DefId.

Fields

def_id: DefId

is_thread_local: bool

ConstRef

A temporary created that references the const with the given DefId

Fields

def_id: DefId

AggregateTemp

A temporary created during the creation of an aggregate (e.g. a temporary for foo in MyStruct { my_field: foo })

BlockTailTemp(BlockTailInfo)

A temporary created for evaluation of some subexpression of some blockโs tail expression (with no intervening statement context).

IfThenRescopeTemp

A temporary created during evaluating if predicate, possibly for pattern matching for lets, and subject to Edition 2024 temporary lifetime rules

Fields

if_then: HirId

DerefTemp

A temporary created during the pass Derefer to avoid itโs retagging

FakeBorrow

A temporary created for borrow checking.

Boring

A local without anything interesting about it.