flowistry_pdg::rustc::mir
Enum Rvalue

pub enum Rvalue<'tcx> {

    Use(Operand<'tcx>),
    Repeat(Operand<'tcx>, Const<'tcx>),
    Ref(Region<'tcx>, BorrowKind, Place<'tcx>),
    ThreadLocalRef(DefId),
    RawPtr(Mutability, Place<'tcx>),
    Len(Place<'tcx>),
    Cast(CastKind, Operand<'tcx>, Ty<'tcx>),
    BinaryOp(BinOp, Box<(Operand<'tcx>, Operand<'tcx>)>),
    NullaryOp(NullOp<'tcx>, Ty<'tcx>),
    UnaryOp(UnOp, Operand<'tcx>),
    Discriminant(Place<'tcx>),
    Aggregate(Box<AggregateKind<'tcx>>, IndexVec<FieldIdx, Operand<'tcx>>),
    ShallowInitBox(Operand<'tcx>, Ty<'tcx>),
    CopyForDeref(Place<'tcx>),

}

The various kinds of rvalues that can appear in MIR.

Not all of these are allowed at every MirPhase - when this is the case, itโs stated below.

Computing any rvalue begins by evaluating the places and operands in some order (Needs clarification: Which order?). These are then used to produce a โvalueโ - the same kind of value that an Operand produces.

Variants

Use(Operand<'tcx>)

Yields the operand unchanged

Repeat(Operand<'tcx>, Const<'tcx>)

Creates an array where each element is the value of the operand.

This is the cause of a bug in the case where the repetition count is zero because the value is not dropped, see #74836.

Corresponds to source code like [x; 32].

Ref(Region<'tcx>, BorrowKind, Place<'tcx>)

Creates a reference of the indicated kind to the place.

There is not much to document here, because besides the obvious parts the semantics of this are essentially entirely a part of the aliasing model. There are many UCG issues discussing exactly what the behavior of this operation should be.

Shallow borrows are disallowed after drop lowering.

ThreadLocalRef(DefId)

Creates a pointer/reference to the given thread local.

The yielded type is a *mut T if the static is mutable, otherwise if the static is extern a *const T, and if neither of those apply a &T.

Note: This is a runtime operation that actually executes code and is in this sense more like a function call. Also, eliminating dead stores of this rvalue causes fn main() {} to SIGILL for some reason that I (JakobDegen) never got a chance to look into.

Needs clarification: Are there weird additional semantics here related to the runtime nature of this operation?

RawPtr(Mutability, Place<'tcx>)

Creates a raw pointer with the indicated mutability to the place.

This is generated by pointer casts like &v as *const _ or raw borrow expressions like &raw const v.

Like with references, the semantics of this operation are heavily dependent on the aliasing model.

Len(Place<'tcx>)

Yields the length of the place, as a usize.

If the type of the place is an array, this is the array length. For slices ([T], not &[T]) this accesses the placeโs metadata to determine the length. This rvalue is ill-formed for places of other types.

This cannot be a UnOp(PtrMetadata, _) because that expects a value, and we only have a place, and UnOp(PtrMetadata, RawPtr(place)) is not a thing.

Cast(CastKind, Operand<'tcx>, Ty<'tcx>)

Performs essentially all of the casts that can be performed via as.

This allows for casts from/to a variety of types.

FIXME: Document exactly which CastKinds allow which types of casts.

BinaryOp(BinOp, Box<(Operand<'tcx>, Operand<'tcx>)>)

Offset has the same semantics as offset, except that the second parameter may be a usize as well.

The comparison operations accept bools, chars, signed or unsigned integers, floats, raw pointers, or function pointers and return a bool. The types of the operands must be matching, up to the usual caveat of the lifetimes in function pointers.

Left and right shift operations accept signed or unsigned integers not necessarily of the same type and return a value of the same type as their LHS. Like in Rust, the RHS is truncated as needed.

The Bit* operations accept signed integers, unsigned integers, or bools with matching types and return a value of that type.

The FooWithOverflow are like the Foo, but returning (T, bool) instead of just T, where the bool is true if the result is not equal to the infinite-precision result.

The remaining operations accept signed integers, unsigned integers, or floats with matching types and return a value of that type.

NullaryOp(NullOp<'tcx>, Ty<'tcx>)

Computes a value as described by the operation.

UnaryOp(UnOp, Operand<'tcx>)

Exactly like BinaryOp, but less operands.

Also does twoโs-complement arithmetic. Negation requires a signed integer or a float; bitwise not requires a signed integer, unsigned integer, or bool. Both operation kinds return a value with the same type as their operand.

Discriminant(Place<'tcx>)

Computes the discriminant of the place, returning it as an integer of type discriminant_ty. Returns zero for types without discriminant.

The validity requirements for the underlying value are undecided for this rvalue, see #91095. Note too that the value of the discriminant is not the same thing as the variant index; use discriminant_for_variant to convert.

Aggregate(Box<AggregateKind<'tcx>>, IndexVec<FieldIdx, Operand<'tcx>>)

Creates an aggregate value, like a tuple or struct.

This is needed because dataflow analysis needs to distinguish dest = Foo { x: ..., y: ... } from dest.x = ...; dest.y = ...; in the case that Foo has a destructor.

Disallowed after deaggregation for all aggregate kinds except Array and Coroutine. After coroutine lowering, Coroutine aggregate kinds are disallowed too.

ShallowInitBox(Operand<'tcx>, Ty<'tcx>)

Transmutes a *mut u8 into shallow-initialized Box<T>.

This is different from a normal transmute because dataflow analysis will treat the box as initialized but its content as uninitialized. Like other pointer casts, this in general affects alias analysis.

CopyForDeref(Place<'tcx>)

A CopyForDeref is equivalent to a read from a place at the codegen level, but is treated specially by drop elaboration. When such a read happens, it is guaranteed (via nature of the mir_opt Derefer in rustc_mir_transform/src/deref_separator) that the only use of the returned value is a deref operation, immediately followed by one or more projections. Drop elaboration treats this rvalue as if the read never happened and just projects further. This allows simplifying various MIR optimizations and codegen backends that previously had to handle deref operations anywhere in a place.