flowistry_pdg::rustc::mir::interpret
Enum ValidationErrorKind

pub enum ValidationErrorKind<'tcx> {

    PointerAsInt {
        expected: ExpectedKind,
    },
    PartialPointer,
    PtrToUninhabited {
        ptr_kind: PointerKind,
        ty: Ty<'tcx>,
    },
    ConstRefToMutable,
    ConstRefToExtern,
    MutableRefToImmutable,
    UnsafeCellInImmutable,
    NullFnPtr,
    NeverVal,
    NullablePtrOutOfRange {
        range: WrappingRange,
        max_value: u128,
    },
    PtrOutOfRange {
        range: WrappingRange,
        max_value: u128,
    },
    OutOfRange {
        value: String,
        range: WrappingRange,
        max_value: u128,
    },
    UninhabitedVal {
        ty: Ty<'tcx>,
    },
    InvalidEnumTag {
        value: String,
    },
    UninhabitedEnumVariant,
    Uninit {
        expected: ExpectedKind,
    },
    InvalidVTablePtr {
        value: String,
    },
    InvalidMetaWrongTrait {
        vtable_dyn_type: &'tcx RawList<(), Binder<TyCtxt<'tcx>, ExistentialPredicate<TyCtxt<'tcx>>>>,
        expected_dyn_type: &'tcx RawList<(), Binder<TyCtxt<'tcx>, ExistentialPredicate<TyCtxt<'tcx>>>>,
    },
    InvalidMetaSliceTooLarge {
        ptr_kind: PointerKind,
    },
    InvalidMetaTooLarge {
        ptr_kind: PointerKind,
    },
    UnalignedPtr {
        ptr_kind: PointerKind,
        required_bytes: u64,
        found_bytes: u64,
    },
    NullPtr {
        ptr_kind: PointerKind,
    },
    DanglingPtrNoProvenance {
        ptr_kind: PointerKind,
        pointer: String,
    },
    DanglingPtrOutOfBounds {
        ptr_kind: PointerKind,
    },
    DanglingPtrUseAfterFree {
        ptr_kind: PointerKind,
    },
    InvalidBool {
        value: String,
    },
    InvalidChar {
        value: String,
    },
    InvalidFnPtr {
        value: String,
    },

}

Variants

PointerAsInt

Fields

expected: ExpectedKind

PartialPointer

PtrToUninhabited

Fields

ptr_kind: PointerKind

ty: Ty<'tcx>

ConstRefToMutable

ConstRefToExtern

MutableRefToImmutable

UnsafeCellInImmutable

NullFnPtr

NeverVal

NullablePtrOutOfRange

Fields

range: WrappingRange

max_value: u128

PtrOutOfRange

Fields

range: WrappingRange

max_value: u128

OutOfRange

Fields

value: String

range: WrappingRange

max_value: u128

UninhabitedVal

Fields

ty: Ty<'tcx>

InvalidEnumTag

Fields

value: String

UninhabitedEnumVariant

Uninit

Fields

expected: ExpectedKind

InvalidVTablePtr

Fields

value: String

InvalidMetaWrongTrait

Fields

vtable_dyn_type: &'tcx RawList<(), Binder<TyCtxt<'tcx>, ExistentialPredicate<TyCtxt<'tcx>>>>

The vtable that was actually referenced by the wide pointer metadata.

expected_dyn_type: &'tcx RawList<(), Binder<TyCtxt<'tcx>, ExistentialPredicate<TyCtxt<'tcx>>>>

The vtable that was expected at the point in MIR that it was accessed.

InvalidMetaSliceTooLarge

Fields

ptr_kind: PointerKind

InvalidMetaTooLarge

Fields

ptr_kind: PointerKind

UnalignedPtr

Fields

ptr_kind: PointerKind

required_bytes: u64

found_bytes: u64

NullPtr

Fields

ptr_kind: PointerKind

DanglingPtrNoProvenance

Fields

ptr_kind: PointerKind

pointer: String

DanglingPtrOutOfBounds

Fields

ptr_kind: PointerKind

DanglingPtrUseAfterFree

Fields

ptr_kind: PointerKind

InvalidBool

Fields

value: String

InvalidChar

Fields

value: String

InvalidFnPtr

Fields

value: String