Trait flowistry_pdg::rustc::mir::interpret::MachineStopType

pub trait MachineStopType: Any + Debug + Send {
    // Required methods
    fn diagnostic_message(&self) -> DiagnosticMessage;
    fn add_args(
        self: Box<Self, Global>,
        adder: &mut dyn FnMut(Cow<'static, str>, DiagnosticArgValue<'static>)
    );
}

A trait for machine-specific errors (or other “machine stop” conditions).

Required Methods

fn diagnostic_message(&self) -> DiagnosticMessage

The diagnostic message for this error

fn add_args( self: Box<Self, Global>, adder: &mut dyn FnMut(Cow<'static, str>, DiagnosticArgValue<'static>) )

Add diagnostic arguments by passing name and value pairs to adder, which are passed to fluent for formatting the translated diagnostic message.