Struct flowistry_pdg::rustc::mir::Body

pub struct Body<'tcx> {

    pub basic_blocks: BasicBlocks<'tcx>,
    pub phase: MirPhase,
    pub pass_count: usize,
    pub source: MirSource<'tcx>,
    pub source_scopes: IndexVec<SourceScope, SourceScopeData<'tcx>>,
    pub generator: Option<Box<GeneratorInfo<'tcx>, Global>>,
    pub local_decls: IndexVec<Local, LocalDecl<'tcx>>,
    pub user_type_annotations: IndexVec<UserTypeAnnotationIndex, CanonicalUserTypeAnnotation<'tcx>>,
    pub arg_count: usize,
    pub spread_arg: Option<Local>,
    pub var_debug_info: Vec<VarDebugInfo<'tcx>, Global>,
    pub span: Span,
    pub required_consts: Vec<Constant<'tcx>, Global>,
    pub is_polymorphic: bool,
    pub injection_phase: Option<MirPhase>,
    pub tainted_by_errors: Option<ErrorGuaranteed>,

}

The lowered representation of a single function.

Fields
basic_blocks: BasicBlocks<'tcx>

A list of basic blocks. References to basic block use a newtyped index type BasicBlock that indexes into this vector.

phase: MirPhase

Records how far through the “desugaring and optimization” process this particular MIR has traversed. This is particularly useful when inlining, since in that context we instantiate the promoted constants and add them to our promoted vector – but those promoted items have already been optimized, whereas ours have not. This field allows us to see the difference and forego optimization on the inlined promoted items.

pass_count: usize

How many passses we have executed since starting the current phase. Used for debug output.

source: MirSource<'tcx>source_scopes: IndexVec<SourceScope, SourceScopeData<'tcx>>

A list of source scopes; these are referenced by statements and used for debuginfo. Indexed by a SourceScope.

generator: Option<Box<GeneratorInfo<'tcx>, Global>>local_decls: IndexVec<Local, LocalDecl<'tcx>>

Declarations of locals.

The first local is the return value pointer, followed by arg_count locals for the function arguments, followed by any user-declared variables and temporaries.

user_type_annotations: IndexVec<UserTypeAnnotationIndex, CanonicalUserTypeAnnotation<'tcx>>

User type annotations.

arg_count: usize

The number of arguments this function takes.

Starting at local 1, arg_count locals will be provided by the caller and can be assumed to be initialized.

If this MIR was built for a constant, this will be 0.

spread_arg: Option<Local>

Mark an argument local (which must be a tuple) as getting passed as its individual components at the LLVM level.

This is used for the “rust-call” ABI.

var_debug_info: Vec<VarDebugInfo<'tcx>, Global>

Debug information pertaining to user variables, including captures.

span: Span

A span representing this MIR, for error reporting.

required_consts: Vec<Constant<'tcx>, Global>

Constants that are required to evaluate successfully for this MIR to be well-formed. We hold in this field all the constants we are not able to evaluate yet.

is_polymorphic: bool

Does this body use generic parameters. This is used for the ConstEvaluatable check.

Note that this does not actually mean that this body is not computable right now. The repeat count in the following example is polymorphic, but can still be evaluated without knowing anything about the type parameter T.

fn test<T>() {
    let _ = [0; std::mem::size_of::<*mut T>()];
}

WARNING: Do not change this flags after the MIR was originally created, even if an optimization removed the last mention of all generic params. We do not want to rely on optimizations and potentially allow things like [u8; std::mem::size_of::<T>() * 0] due to this.

injection_phase: Option<MirPhase>

The phase at which this MIR should be “injected” into the compilation process.

Everything that comes before this MirPhase should be skipped.

This is only Some if the function that this body comes from was annotated with rustc_custom_mir.

tainted_by_errors: Option<ErrorGuaranteed>