Module flowistry_pdg::rustc_proxies

Proxies for Rustc types used within the PDG.

Each type has an identical set of fields to the corresponding Rustc type. Paralegal serializes the PDG into these types, which are read by downstream property checkers.

Macros

proxy_index 🔒

Generates a struct that is a proxy for a Rustc index type.

proxy_struct 🔒

Generates a struct that is a proxy for a Rustc type.

Structs

BasicBlock

Proxy for mir::BasicBlock

BodyId

Proxy for hir::BodyId

CrateNum

Proxy for hir::def_id::CrateNum

DefId

Proxy for def_id::DefId

DefIndex

Proxy for def_id::DefIndex

HirId

Proxy for hir::HirId

ItemLocalId

Proxy for hir::ItemLocalId

LocalDefId

Proxy for def_id::LocalDefId

Location

Proxy for mir::Location

OwnerId

Proxy for hir_id::OwnerHid