flowistry_pdg::rustc::mir
Enum StatementKind

pub enum StatementKind<'tcx> {

    Assign(Box<(Place<'tcx>, Rvalue<'tcx>)>),
    FakeRead(Box<(FakeReadCause, Place<'tcx>)>),
    SetDiscriminant {
        place: Box<Place<'tcx>>,
        variant_index: VariantIdx,
    },
    Deinit(Box<Place<'tcx>>),
    StorageLive(Local),
    StorageDead(Local),
    Retag(RetagKind, Box<Place<'tcx>>),
    PlaceMention(Box<Place<'tcx>>),
    AscribeUserType(Box<(Place<'tcx>, UserTypeProjection)>, Variance),
    Coverage(CoverageKind),
    Intrinsic(Box<NonDivergingIntrinsic<'tcx>>),
    ConstEvalCounter,
    Nop,
    BackwardIncompatibleDropHint {
        place: Box<Place<'tcx>>,
        reason: BackwardIncompatibleDropReason,
    },

}

The various kinds of statements that can appear in MIR.

Not all of these are allowed at every MirPhase. Check the documentation there to see which ones you do not have to worry about. The MIR validator will generally enforce such restrictions, causing an ICE if they are violated.

Variants

Assign(Box<(Place<'tcx>, Rvalue<'tcx>)>)

Assign statements roughly correspond to an assignment in Rust proper (x = ...) except without the possibility of dropping the previous value (that must be done separately, if at all). The exact way this works is undecided. It probably does something like evaluating the LHS to a place and the RHS to a value, and then storing the value to the place. Various parts of this may do type specific things that are more complicated than simply copying bytes.

Needs clarification: The implication of the above idea would be that assignment implies that the resulting value is initialized. I believe we could commit to this separately from committing to whatever part of the memory model we would need to decide on to make the above paragraph precise. Do we want to?

Assignments in which the types of the place and rvalue differ are not well-formed.

Needs clarification: Do we ever want to worry about non-free (in the body) lifetimes for the typing requirement in post drop-elaboration MIR? I think probably not - I’m not sure we could meaningfully require this anyway. How about free lifetimes? Is ignoring this interesting for optimizations? Do we want to allow such optimizations?

Needs clarification: We currently require that the LHS place not overlap with any place read as part of computation of the RHS for some rvalues (generally those not producing primitives). This requirement is under discussion in #68364. As a part of this discussion, it is also unclear in what order the components are evaluated.

See Rvalue documentation for details on each of those.

FakeRead(Box<(FakeReadCause, Place<'tcx>)>)

This represents all the reading that a pattern match may do (e.g., inspecting constants and discriminant values), and the kind of pattern it comes from. This is in order to adapt potential error messages to these specific patterns.

Note that this also is emitted for regular let bindings to ensure that locals that are never accessed still get some sanity checks for, e.g., let x: ! = ..;

When executed at runtime this is a nop.

Disallowed after drop elaboration.

SetDiscriminant

Write the discriminant for a variant to the enum Place.

This is permitted for both coroutines and ADTs. This does not necessarily write to the entire place; instead, it writes to the minimum set of bytes as required by the layout for the type.

Fields

place: Box<Place<'tcx>>

variant_index: VariantIdx

Deinit(Box<Place<'tcx>>)

Deinitializes the place.

This writes uninit bytes to the entire place.

StorageLive(Local)

StorageLive and StorageDead statements mark the live range of a local.

At any point during the execution of a function, each local is either allocated or unallocated. Except as noted below, all locals except function parameters are initially unallocated. StorageLive statements cause memory to be allocated for the local while StorageDead statements cause the memory to be freed. In other words, StorageLive/StorageDead act like the heap operations allocate/deallocate, but for stack-allocated local variables. Using a local in any way (not only reading/writing from it) while it is unallocated is UB.

Some locals have no StorageLive or StorageDead statements within the entire MIR body. These locals are implicitly allocated for the full duration of the function. There is a convenience method at rustc_mir_dataflow::storage::always_storage_live_locals for computing these locals.

If the local is already allocated, calling StorageLive again will implicitly free the local and then allocate fresh uninitilized memory. If a local is already deallocated, calling StorageDead again is a NOP.

StorageDead(Local)

See StorageLive above.

Retag(RetagKind, Box<Place<'tcx>>)

Retag references in the given place, ensuring they got fresh tags.

This is part of the Stacked Borrows model. These statements are currently only interpreted by miri and only generated when -Z mir-emit-retag is passed. See https://internals.rust-lang.org/t/stacked-borrows-an-aliasing-model-for-rust/8153/ for more details.

For code that is not specific to stacked borrows, you should consider retags to read and modify the place in an opaque way.

Only RetagKind::Default and RetagKind::FnEntry are permitted.

PlaceMention(Box<Place<'tcx>>)

This statement exists to preserve a trace of a scrutinee matched against a wildcard binding. This is especially useful for let _ = PLACE; bindings that desugar to a single PlaceMention(PLACE).

When executed at runtime, this computes the given place, but then discards it without doing a load. let _ = *ptr; is fine even if the pointer is dangling.

AscribeUserType(Box<(Place<'tcx>, UserTypeProjection)>, Variance)

Encodes a user’s type ascription. These need to be preserved intact so that NLL can respect them. For example:

let a: T = y;

The effect of this annotation is to relate the type T_y of the place y to the user-given type T. The effect depends on the specified variance:

Covariant – requires that T_y <: T

Contravariant – requires that T_y :> T

Invariant – requires that T_y == T

Bivariant – no effect

When executed at runtime this is a nop.

Disallowed after drop elaboration.

Coverage(CoverageKind)

Carries control-flow-sensitive information injected by -Cinstrument-coverage, such as where to generate physical coverage-counter-increments during codegen.

Coverage statements are used in conjunction with the coverage mappings and other information stored in the function’s mir::Body::function_coverage_info. (For inlined MIR, take care to look up the original function’s coverage info.)

Interpreters and codegen backends that don’t support coverage instrumentation can usually treat this as a no-op.

Intrinsic(Box<NonDivergingIntrinsic<'tcx>>)

Denotes a call to an intrinsic that does not require an unwind path and always returns. This avoids adding a new block and a terminator for simple intrinsics.

ConstEvalCounter

Instructs the const eval interpreter to increment a counter; this counter is used to track how many steps the interpreter has taken. It is used to prevent the user from writing const code that runs for too long or infinitely. Other than in the const eval interpreter, this is a no-op.

Nop

No-op. Useful for deleting instructions without affecting statement indices.

BackwardIncompatibleDropHint

Marker statement indicating where place would be dropped. This is semantically equivalent to Nop, so codegen and MIRI should interpret this statement as such. The only use case of this statement is for linting in MIR to detect temporary lifetime changes.

Fields

place: Box<Place<'tcx>>

Place to drop

reason: BackwardIncompatibleDropReason

Reason for backward incompatibility