flowistry_pdg::rustc::mir::interpret
Enum UndefinedBehaviorInfo

pub enum UndefinedBehaviorInfo<'tcx> {

    Ub(String),
    Custom(CustomSubdiagnostic<'tcx>),
    ValidationError(ValidationErrorInfo<'tcx>),
    Unreachable,
    BoundsCheckFailed {
        len: u64,
        index: u64,
    },
    DivisionByZero,
    RemainderByZero,
    DivisionOverflow,
    RemainderOverflow,
    PointerArithOverflow,
    ArithOverflow {
        intrinsic: Symbol,
    },
    ShiftOverflow {
        intrinsic: Symbol,
        shift_amount: Either<u128, i128>,
    },
    InvalidMeta(InvalidMetaKind),
    UnterminatedCString(Pointer<AllocId>),
    PointerUseAfterFree(AllocId, CheckInAllocMsg),
    PointerOutOfBounds {
        alloc_id: AllocId,
        alloc_size: Size,
        ptr_offset: i64,
        inbounds_size: i64,
        msg: CheckInAllocMsg,
    },
    DanglingIntPointer {
        addr: u64,
        inbounds_size: i64,
        msg: CheckInAllocMsg,
    },
    AlignmentCheckFailed(Misalignment, CheckAlignMsg),
    WriteToReadOnly(AllocId),
    DerefFunctionPointer(AllocId),
    DerefVTablePointer(AllocId),
    InvalidBool(u8),
    InvalidChar(u32),
    InvalidTag(Scalar<AllocId>),
    InvalidFunctionPointer(Pointer<AllocId>),
    InvalidVTablePointer(Pointer<AllocId>),
    InvalidVTableTrait {
        vtable_dyn_type: &'tcx RawList<(), Binder<TyCtxt<'tcx>, ExistentialPredicate<TyCtxt<'tcx>>>>,
        expected_dyn_type: &'tcx RawList<(), Binder<TyCtxt<'tcx>, ExistentialPredicate<TyCtxt<'tcx>>>>,
    },
    InvalidStr(Utf8Error),
    InvalidUninitBytes(Option<(AllocId, BadBytesAccess)>),
    DeadLocal,
    ScalarSizeMismatch(ScalarSizeMismatch),
    UninhabitedEnumVariantWritten(VariantIdx),
    UninhabitedEnumVariantRead(VariantIdx),
    InvalidNichedEnumVariantWritten {
        enum_ty: Ty<'tcx>,
    },
    AbiMismatchArgument {
        caller_ty: Ty<'tcx>,
        callee_ty: Ty<'tcx>,
    },
    AbiMismatchReturn {
        caller_ty: Ty<'tcx>,
        callee_ty: Ty<'tcx>,
    },

}

Error information for when the program caused Undefined Behavior.

Variants

Ub(String)

Free-form case. Only for errors that are never caught! Used by miri

Custom(CustomSubdiagnostic<'tcx>)

A custom (free-form) fluent-translated error, created by err_ub_custom!.

ValidationError(ValidationErrorInfo<'tcx>)

Validation error.

Unreachable

Unreachable code was executed.

BoundsCheckFailed

A slice/array index projection went out-of-bounds.

Fields

len: u64

index: u64

DivisionByZero

Something was divided by 0 (x / 0).

RemainderByZero

Something was “remainded” by 0 (x % 0).

DivisionOverflow

Signed division overflowed (INT_MIN / -1).

RemainderOverflow

Signed remainder overflowed (INT_MIN % -1).

PointerArithOverflow

Overflowing inbounds pointer arithmetic.

ArithOverflow

Overflow in arithmetic that may not overflow.

Fields

intrinsic: Symbol

ShiftOverflow

Shift by too much.

Fields

intrinsic: Symbol

shift_amount: Either<u128, i128>

InvalidMeta(InvalidMetaKind)

Invalid metadata in a wide pointer

UnterminatedCString(Pointer<AllocId>)

Reading a C string that does not end within its allocation.

PointerUseAfterFree(AllocId, CheckInAllocMsg)

Using a pointer after it got freed.

PointerOutOfBounds

Used a pointer outside the bounds it is valid for.

Fields

alloc_id: AllocId

alloc_size: Size

ptr_offset: i64

inbounds_size: i64

The size of the memory range that was expected to be in-bounds.

msg: CheckInAllocMsg

DanglingIntPointer

Using an integer as a pointer in the wrong way.

Fields

addr: u64

inbounds_size: i64

The size of the memory range that was expected to be in-bounds (or 0 if we need an allocation but not any actual memory there, e.g. for function pointers).

msg: CheckInAllocMsg

AlignmentCheckFailed(Misalignment, CheckAlignMsg)

Used a pointer with bad alignment.

WriteToReadOnly(AllocId)

Writing to read-only memory.

DerefFunctionPointer(AllocId)

Trying to access the data behind a function pointer.

DerefVTablePointer(AllocId)

Trying to access the data behind a vtable pointer.

InvalidBool(u8)

Using a non-boolean u8 as bool.

InvalidChar(u32)

Using a non-character u32 as character.

InvalidTag(Scalar<AllocId>)

The tag of an enum does not encode an actual discriminant.

InvalidFunctionPointer(Pointer<AllocId>)

Using a pointer-not-to-a-function as function pointer.

InvalidVTablePointer(Pointer<AllocId>)

Using a pointer-not-to-a-vtable as vtable pointer.

InvalidVTableTrait

Using a vtable for the wrong trait.

Fields

vtable_dyn_type: &'tcx RawList<(), Binder<TyCtxt<'tcx>, ExistentialPredicate<TyCtxt<'tcx>>>>

The vtable that was actually referenced by the wide pointer metadata.

expected_dyn_type: &'tcx RawList<(), Binder<TyCtxt<'tcx>, ExistentialPredicate<TyCtxt<'tcx>>>>

The vtable that was expected at the point in MIR that it was accessed.

InvalidStr(Utf8Error)

Using a string that is not valid UTF-8,

InvalidUninitBytes(Option<(AllocId, BadBytesAccess)>)

Using uninitialized data where it is not allowed.

DeadLocal

Working with a local that is not currently live.

ScalarSizeMismatch(ScalarSizeMismatch)

Data size is not equal to target size.

UninhabitedEnumVariantWritten(VariantIdx)

A discriminant of an uninhabited enum variant is written.

UninhabitedEnumVariantRead(VariantIdx)

An uninhabited enum variant is projected.

InvalidNichedEnumVariantWritten

Trying to set discriminant to the niched variant, but the value does not match.

Fields

enum_ty: Ty<'tcx>

AbiMismatchArgument

ABI-incompatible argument types.

Fields

caller_ty: Ty<'tcx>

callee_ty: Ty<'tcx>

AbiMismatchReturn

ABI-incompatible return types.

Fields

caller_ty: Ty<'tcx>

callee_ty: Ty<'tcx>