flowistry_pdg::rustc::mir::interpret
Trait MachineStopType

pub trait MachineStopType:
    Any
    + Debug
    + Send {
    // Required methods
    fn diagnostic_message(&self) -> DiagMessage;
    fn add_args(
        self: Box<Self>,
        adder: &mut dyn FnMut(Cow<'static, str>, DiagArgValue),
    );
}

A trait for machine-specific errors (or other “machine stop” conditions).

Required Methods

fn diagnostic_message(&self) -> DiagMessage

The diagnostic message for this error

fn add_args( self: Box<Self>, adder: &mut dyn FnMut(Cow<'static, str>, DiagArgValue), )

Add diagnostic arguments by passing name and value pairs to adder, which are passed to fluent for formatting the translated diagnostic message.