flowistry_pdg::rustc::mir
Module pretty

Structs

PrettyPrintMirOptions

Cosmetic options for pretty-printing the MIR contents, gathered from the CLI. Each pass can override these when dumping its own specific MIR information with dump_mir_with_options.

Enums

PassWhere

An indication of where we are in the control flow graph. Used for printing extra information in dump_mir

Functions

create_dump_file

Attempts to open a file where we should dump a given MIR or other bit of MIR-related data. Used by mir-dump, but also by other bits of code (e.g., NLL inference) that dump graphviz data or other things, and hence takes the extension as an argument.

display_allocation

Dumps the size and metadata and content of an allocation to the given writer. The expectation is that the caller first prints other relevant metadata, so the exact format of this function is (without leading or trailing newline):

dump_enabled

dump_mir

If the session is properly configured, dumps a human-readable representation of the MIR (with default pretty-printing options) into:

dump_mir_def_ids

dump_mir_with_options

If the session is properly configured, dumps a human-readable representation of the MIR, with the given pretty-printing options.

write_allocation_bytes

The prefix argument allows callers to add an arbitrary prefix before each line (even if there is only one line). Note that your prefix should contain a trailing space as the lines are printed directly after it.

write_allocations

Find all AllocIds mentioned (recursively) in the MIR body and print their corresponding allocations.

write_mir_fn

Write out a human-readable textual representation for the given function.

write_mir_pretty

Write out a human-readable textual representation for the given MIR, with the default PrettyPrintMirOptions.