flowistry_pdg::rustc::mir
Struct Place

pub struct Place<'tcx> {
    pub local: Local,
    pub projection: &'tcx RawList<(), ProjectionElem<Local, Ty<'tcx>>>,
}

Places roughly correspond to a “location in memory.” Places in MIR are the same mathematical object as places in Rust. This of course means that what exactly they are is undecided and part of the Rust memory model. However, they will likely contain at least the following pieces of information in some form:

The address in memory that the place refers to.

The provenance with which the place is being accessed.

The type of the place and an optional variant index. See PlaceTy.

Optionally, some metadata. This exists if and only if the type of the place is not Sized.

We’ll give a description below of how all pieces of the place except for the provenance are calculated. We cannot give a description of the provenance, because that is part of the undecided aliasing model - we only include it here at all to acknowledge its existence.

Each local naturally corresponds to the place Place { local, projection: [] }. This place has the address of the local’s allocation and the type of the local.

Needs clarification: Unsized locals seem to present a bit of an issue. Their allocation can’t actually be created on StorageLive, because it’s unclear how big to make the allocation. Furthermore, MIR produces assignments to unsized locals, although that is not permitted under #![feature(unsized_locals)] in Rust. Besides just putting “unsized locals are special and different” in a bunch of places, I (JakobDegen) don’t know how to incorporate this behavior into the current MIR semantics in a clean way - possibly this needs some design work first.

For places that are not locals, ie they have a non-empty list of projections, we define the values as a function of the parent place, that is the place with its last ProjectionElem stripped. The way this is computed of course depends on the kind of that last projection element:

Downcast: This projection sets the place’s variant index to the given one, and makes no other changes. A Downcast projection must always be followed immediately by a Field projection.

Field: Field projections take their parent place and create a place referring to one of the fields of the type. The resulting address is the parent address, plus the offset of the field. The type becomes the type of the field. If the parent was unsized and so had metadata associated with it, then the metadata is retained if the field is unsized and thrown out if it is sized.

These projections are only legal for tuples, ADTs, closures, and coroutines. If the ADT or coroutine has more than one variant, the parent place’s variant index must be set, indicating which variant is being used. If it has just one variant, the variant index may or may not be included - the single possible variant is inferred if it is not included.

OpaqueCast: This projection changes the place’s type to the given one, and makes no other changes. A OpaqueCast projection on any type other than an opaque type from the current crate is not well-formed.

ConstantIndex: Computes an offset in units of T into the place as described in the documentation for the ProjectionElem. The resulting address is the parent’s address plus that offset, and the type is T. This is only legal if the parent place has type [T; N] or [T] (not &[T]). Since such a T is always sized, any resulting metadata is thrown out.

Subslice: This projection calculates an offset and a new address in a similar manner as ConstantIndex. It is also only legal on [T; N] and [T]. However, this yields a Place of type [T], and additionally sets the metadata to be the length of the subslice.

Index: Like ConstantIndex, only legal on [T; N] or [T]. However, Index additionally takes a local from which the value of the index is computed at runtime. Computing the value of the index involves interpreting the Local as a Place { local, projection: [] }, and then computing its value as if done via Operand::Copy. The array/slice is then indexed with the resulting value. The local must have type usize.

Deref: Derefs are the last type of projection, and the most complicated. They are only legal on parent places that are references, pointers, or Box. A Deref projection begins by loading a value from the parent place, as if by Operand::Copy. It then dereferences the resulting pointer, creating a place of the pointee’s type. The resulting address is the address that was stored in the pointer. If the pointee type is unsized, the pointer additionally stored the value of the metadata.

The “validity invariant” of places is the same as that of raw pointers, meaning that e.g. *ptr on a dangling or unaligned pointer is never UB. (Later doing a load/store on that place or turning it into a reference can be UB though!) The only ways for a place computation can cause UB are:

On a Deref projection, we do an actual load of the inner place, with all the usual consequences (the inner place must be based on an aligned pointer, it must point to allocated memory, the aliasig model must allow reads, this must not be a data race).

For the projections that perform pointer arithmetic, the offset must in-bounds of an allocation (i.e., the preconditions of ptr::offset must be met).

Fields
local: Localprojection: &'tcx RawList<(), ProjectionElem<Local, Ty<'tcx>>>

projection out of a place (access a field, deref a pointer, etc)